inline encoder file into AssetGenerator
