+++ lib/asset/AssetGenerator.js
@@ -1,14 +1,15 @@
 /*
 	MIT License http://www.opensource.org/licenses/mit-license.php
-	Author Yuta Hiroto @hiroppy
+	Author Sergey Melyukov @smelukov
 */
 
 "use strict";
 
+const mime = require("mime");
+const path = require("path");
 const { RawSource } = require("webpack-sources");
 const Generator = require("../Generator");
 const RuntimeGlobals = require("../RuntimeGlobals");
-const encoder = require("./encoder");
 
 /** @typedef {import("webpack-sources").Source} Source */
 /** @typedef {import("../../declarations/plugins/AssetModulesPlugin").AssetModulesPluginOptions} AssetModulesPluginOptions */
@@ -18,6 +19,102 @@ const encoder = require("./encoder");
 /** @typedef {import("../Module")} Module */
 /** @typedef {import("../NormalModule")} NormalModule */
 /** @typedef {import("../RuntimeTemplate")} RuntimeTemplate */
+
+/**
+ * @type {Map<string|Buffer, string|null>}
+ */
+const dataUrlFnCache = new Map();
+
+/**
+ * @param {NormalModule} module the module
+ * @param {AssetModulesPluginOptions} options the options to encode
+ * @returns {boolean} should emit additional asset for the module
+ */
+const shouldEmitAsset = (module, options) => {
+	const originalSource = module.originalSource();
+	if (typeof options.dataUrl === "function") {
+		return (
+			options.dataUrl.call(null, module, module.nameForCondition()) === false
+		);
+	}
+
+	if (options.dataUrl === false) {
+		return true;
+	}
+
+	return originalSource.size() > options.dataUrl.maxSize;
+};
+
+/**
+ * @param {AssetModulesPluginOptions} options the options to the encoder
+ * @returns {AssetModulesPluginOptions} normalized options
+ */
+const prepareOptions = (options = {}) => {
+	const dataUrl = options.dataUrl || {};
+
+	if (options.dataUrl === false) {
+		return {
+			dataUrl: false
+		};
+	}
+
+	if (typeof dataUrl === "function") {
+		return {
+			dataUrl: (source, resourcePath) => {
+				if (dataUrlFnCache.has(source)) {
+					return dataUrlFnCache.get(source);
+				}
+
+				const encoded = dataUrl.call(null, source, resourcePath);
+				dataUrlFnCache.set(source, encoded);
+
+				return encoded;
+			}
+		};
+	}
+
+	return {
+		dataUrl: {
+			encoding: "base64",
+			maxSize: 8192,
+			...dataUrl
+		}
+	};
+};
+
+/**
+ * @param {NormalModule} module a module to encode
+ * @param {AssetModulesPluginOptions} options the options to the encoder
+ * @returns {string|null} encoded source
+ */
+const encode = (module, options) => {
+	if (shouldEmitAsset(module, options)) {
+		return null;
+	}
+
+	const originalSource = module.originalSource();
+	let content = originalSource.source();
+
+	if (typeof options.dataUrl === "function") {
+		return options.dataUrl.call(null, content, module.nameForCondition());
+	}
+
+	// @ts-ignore non-false dataUrl ensures in shouldEmitAsset above
+	const encoding = options.dataUrl.encoding;
+	const extname = path.extname(module.nameForCondition());
+	// @ts-ignore non-false dataUrl ensures in shouldEmitAsset above
+	const mimeType = options.dataUrl.mimetype || mime.getType(extname);
+
+	if (encoding === "base64") {
+		if (typeof content === "string") {
+			content = Buffer.from(content);
+		}
+
+		content = content.toString("base64");
+	}
+
+	return `data:${mimeType}${encoding ? `;${encoding}` : ""},${content}`;
+};
 
 const JS_TYPES = new Set(["javascript"]);
 const JS_AND_ASSET_TYPES = new Set(["javascript", "asset"]);
@@ -30,7 +127,7 @@ class AssetGenerator extends Generator {
 	constructor(compilation, options) {
 		super();
 		this.compilation = compilation;
-		this.options = encoder.prepareOptions(options);
+		this.options = prepareOptions(options);
 	}
 
 	/**
@@ -45,8 +142,8 @@ class AssetGenerator extends Generator {
 
 		runtimeRequirements.add(RuntimeGlobals.module);
 
-		if (!encoder.shouldEmitAsset(module, this.options)) {
-			const encodedSource = encoder.encode(module, this.options);
+		if (!shouldEmitAsset(module, this.options)) {
+			const encodedSource = encode(module, this.options);
 			return new RawSource(
 				`${RuntimeGlobals.module}.exports = ${JSON.stringify(encodedSource)};`
 			);
@@ -75,7 +172,7 @@ class AssetGenerator extends Generator {
 	 * @returns {Set<string>} available types (do not mutate)
 	 */
 	getTypes(module) {
-		if (encoder.shouldEmitAsset(module, this.options)) {
+		if (shouldEmitAsset(module, this.options)) {
 			return JS_AND_ASSET_TYPES;
 		}
 
@@ -98,7 +195,7 @@ class AssetGenerator extends Generator {
 			return originalSource.size();
 		}
 
-		if (encoder.shouldEmitAsset(module, this.options)) {
+		if (shouldEmitAsset(module, this.options)) {
 			// it's only estimated so this number is probably fine
 			// Example: m.exports=r.p+"0123456789012345678901.ext"
 			return 42;
